Update DOCS for Balancing Add Commnets

+++ Assets/1.Main/2.Script/Controller/Object/Generator.cs
@@ -62,7 +62,7 @@ public class Generator : BuildableObject
         if (maxUpgradeDefence > 5) return;
         m_stat.Defence += defence;
     }
-    public void IncreaseHPRegen(int regen)
+    public void IncreaseHPRegen(float regen)
     {
         maxUpgradeRegen++;
         if (maxUpgradeRegen > 5) return;
